[docs] refactor: group training feature examples
