+++ examples/training_features/README.md
@@ -1,0 +1,13 @@
+# Training Feature Examples
+
+This directory groups examples that focus on training-system features rather than model families or checkpoint conversion.
+
+## Contents
+
+| Directory | Focus |
+|-----------|-------|
+| `decentralized_pg/` | Explicit process-group construction and `use_decentralized_pg=True` training runs |
+| `long_context/` | Long-context training launch examples |
+| `resiliency/` | Fault tolerance and straggler detection examples |
+
+Run commands from the repository root so relative paths in the child READMEs and launch scripts resolve correctly.
